requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cleanup==5.1.0
django-compat==1.0.15
django-cors-headers==3.7.0
django-crontab==0.7.1
django-debug-toolbar==3.2
djangorestframework==3.12.4
gunicorn==20.1.0
persiantools==2.1.1
Pillow==8.1.2
pytz==2021.1
sqlparse==0.4.1
validate-email==1.3
whitenoise==5.2.0
channels>=2.1.5
channels-redis>=2.3.1
daphne
